Generate paths using special $package$ field.
