+++ src/main/g8/default.properties
@@ -3,4 +3,4 @@ name=CHANGE_ME
 project_description=CHANGE_ME
 organisation_domain=net
 organisation=cakesolutions
-package_structure=$organisation_domain$.$organisation$.$name$
+package=$organisation_domain$.$organisation$.$name$
