+++ src/main/g8/library/core/src/main/scala/$package$/core/application/workflow/RetryExceededException.scala
@@ -1,0 +1,15 @@
+package $package$.core.application.workflow
+
+/**
+  * Exception message indicating that a task's retry logic has failed.
+  *
+  * @param message @see java.lang.Exception
+  */
+final class RetryExceededException(message: String)
+    extends Exception(message) {
+
+  /** @see java.lang.Exception */
+  override def toString: String = {
+    s"RetryExceededException(\$message)"
+  }
+}
