+++ src/main/g8/library/core/src/main/scala/$package$/core/utils/ValueDiscard.scala
