+++ src/main/g8/mains/server/src/test/scala/$package$/templating/CakeNewCommandSpec.scala
